refactor: rename config.ts => consts.ts
Taken from starter blog, feels better.
