+++ src/components/BaseHead.astro
@@ -1,5 +1,5 @@
 ---
-import { SITE_TITLE } from "../config";
+import { SITE_TITLE } from "../consts";
 // Import the global.css file here so that it is included on
 // all pages through the use of the <BaseHead /> component.
 import "../styles/global.css";
