+++ src/components/Header.astro
@@ -1,6 +1,6 @@
 ---
 import HeaderLink from "./HeaderLink.astro";
-import { SITE_TITLE } from "../config";
+import { SITE_TITLE } from "../consts";
 ---
 
 <header>
