+++ src/pages/notes/index.astro
@@ -2,7 +2,7 @@
 import BaseHead from "../../components/BaseHead.astro";
 import Header from "../../components/Header.astro";
 import Footer from "../../components/Footer.astro";
-import { SITE_TITLE, SITE_DESCRIPTION } from "../../config";
+import { SITE_TITLE, SITE_DESCRIPTION } from "../../consts";
 
 // Use Astro.glob() to fetch all posts, and then sort them by date.
 const posts = (await Astro.glob("./**/*.{md,mdx}")).sort(
